refactor: autocompletion code
